removed bot id from feedback
